reduce complexity
