+++ src/includes/gateways/standard/class-alma-payment-gateway-pay-now.php
@@ -46,15 +46,4 @@ class Alma_Payment_Gateway_Pay_Now extends Alma_Payment_Gateway_Standard {
 	public function get_gateway_id() {
 		return Alma_Constants_Helper::GATEWAY_ID_PAY_NOW;
 	}
-
-	/**
-	 * Has fields.
-	 *
-	 * @return true
-	 */
-	public function has_fields() {
-		return true;
-	}
-
-
 } // end \Alma_Pay_Gateway class
